STCOM-934: fixed problems with shortcut
