+++ lib/AdvancedSearch/AdvancedSearch.js
@@ -7,7 +7,6 @@ import {
 import PropTypes from 'prop-types';
 import { useIntl } from 'react-intl';
 
-import { HotKeys } from '../HotKeys';
 import AdvancedSearchModal from './components/AdvancedSearchModal';
 import AdvancedSearchRow from './components/AdvancedSearchRow';
 import defaultQueryBuilder from './utilities/defaultQueryBuilder';
@@ -124,7 +123,9 @@ const AdvancedSearch = ({
       if (row[FIELD_NAMES.SEARCH_OPTION] === QUERY_OPTION_VALUE) {
         // split row with multiple search options into array of rows
 
-        return [...finalRows, ...splitRow(row)];
+        const splittedRow = splitRow(row);
+
+        return [...finalRows, ...(splittedRow.length ? splittedRow : [{ bool:'and', search: '', searchOption: 'keyword' }])];
       }
 
       return [...finalRows, row];
@@ -161,16 +162,8 @@ const AdvancedSearch = ({
     onCancel();
   }, [onCancel]);
 
-  const hotKeys = useMemo(() => ({
-    search: ['enter'],
-  }), []);
-
-  const handlers = useMemo(() => ({
-    search: handleSearch,
-  }), [handleSearch]);
-
   const renderRows = () => {
-    return new Array(ROW_COUNT).fill(null).map((_, index) => (
+    return new Array(rowState.length).fill(null).map((_, index) => (
       <AdvancedSearchRow
         key={`$advanced-search-row-${index}`}
         index={index}
@@ -188,10 +181,7 @@ const AdvancedSearch = ({
   const filledRows = filterFilledRows(rowState);
 
   return (
-    <HotKeys
-      keyMap={hotKeys}
-      handlers={handlers}
-    >
+    <>
       <AdvancedSearchModal
         open={open}
         searchOptions={searchOptionsWithQuery}
@@ -201,7 +191,7 @@ const AdvancedSearch = ({
         {renderRows()}
       </AdvancedSearchModal>
       {children ? children({ resetRows, rowState: filledRows }) : null}
-    </HotKeys>
+    </>
   );
 };
