+++ lib/AdvancedSearch/components/AdvancedSearchModal/AdvancedSearchModal.js
@@ -2,6 +2,7 @@ import { useMemo } from 'react';
 import PropTypes from 'prop-types';
 import { useIntl } from 'react-intl';
 
+import { HotKeys } from '../../../HotKeys';
 import Modal from '../../../Modal';
 import ModalFooter from '../../../ModalFooter';
 import Button from '../../../Button';
@@ -52,6 +53,14 @@ const AdvancedSearchModal = ({
     );
   }, [onSearch, onCancel, cancelLabel, searchLabel]);
 
+  const hotKeys = useMemo(() => ({
+    search: ['enter'],
+  }), []);
+
+  const handlers = useMemo(() => ({
+    search: onSearch,
+  }), [onSearch]);
+
   return (
     <Modal
       open={open}
@@ -63,7 +72,12 @@ const AdvancedSearchModal = ({
       footer={footer}
       aria-label={modalTitle}
     >
-      {children}
+      <HotKeys
+        keyMap={hotKeys}
+        handlers={handlers}
+      >
+        {children}
+      </HotKeys>
     </Modal>
   );
 };
